Added dockerfiles for each micro service and also organised each service into corresponding packages
