+++ src/main/java/Dockerfile
@@ -1,0 +1,17 @@
+# Use a Node.js base image
+FROM node:14
+
+# Create app directory
+WORKDIR /usr/src/app
+
+# Install app dependencies
+# A wildcard is used to ensure both package.json AND package-lock.json are copied
+COPY package*.json ./
+
+RUN npm install
+
+# Bundle app source
+COPY . .
+
+EXPOSE 3000
+CMD ["node", "server.js"]
