--- src/main/java/ie/atu/project/cicd_project/MemberRepo.java
@@ -1,7 +1,0 @@
-package ie.atu.project.cicd_project;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepo extends JpaRepository< Member, Long> {
-
-}
