+++ src/main/java/ie/atu/project/cicd_project/MemeberMicroService/Member.java
@@ -1,4 +1,4 @@
-package ie.atu.project.cicd_project;
+package ie.atu.project.cicd_project.MemeberMicroService;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
