+++ src/main/java/ie/atu/project/cicd_project/MemeberMicroService/MemberRepo.java
@@ -1,0 +1,7 @@
+package ie.atu.project.cicd_project.MemeberMicroService;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepo extends JpaRepository<Member, Long> {
+
+}
