+++ src/main/java/ie/atu/project/cicd_project/TrainingSessionMicroService/TrainingSessionRepo.java
@@ -1,4 +1,4 @@
-package ie.atu.project.cicd_project;
+package ie.atu.project.cicd_project.TrainingSessionMicroService;
 
 import org.springframework.data.jpa.repository.JpaRepository;
