+++ src/main/java/ie/atu/project/cicd_project/docker-compose.yml
@@ -1,0 +1,63 @@
+version: '3.8'
+
+services:
+  event-service:
+    build:
+      context: ./ie.atu.project.cicd_project/EventMicroService
+      dockerfile: Dockerfile
+    ports:
+      - "8080:8080"
+
+  inventory-service:
+    build:
+      context: ./ie.atu.project.cicd_project/InventoryMicroService
+      dockerfile: Dockerfile
+    ports:
+      - "8081:8080"
+
+  member-service:
+    build:
+      context: ./ie.atu.project.cicd_project/MemberMicroService
+      dockerfile: Dockerfile
+    ports:
+      - "8082:8080"
+
+  ranking-service:
+    build:
+      context: ./ie.atu.project.cicd_project/RankingMicroService
+      dockerfile: Dockerfile
+    ports:
+      - "8083:8080"
+
+  training-session-service:
+    build:
+      context: ./ie.atu.project.cicd_project/TrainingSessionMicroService
+      dockerfile: Dockerfile
+    ports:
+      - "8084:8080"
+
+  mysql-database:
+    image: mysql:latest
+    environment:
+      MYSQL_DATABASE: 'mydatabase'
+      MYSQL_USER: 'user'
+      MYSQL_PASSWORD: 'password'
+      MYSQL_ROOT_PASSWORD: 'password'
+    ports:
+      - "3306:3306"
+    volumes:
+      - db-data:/var/lib/mysql
+
+  h2-database:
+    image: oscarfonts/h2
+    ports:
+      - "1521:1521"
+      - "81:81"
+    volumes:
+      - h2-data:/h2-data
+    environment:
+      - H2_OPTIONS=-ifNotExists
+
+volumes:
+  db-data:
+  h2-data:
